Remove the GET route and return the captured parameters in the POST response.
