+++ .gitattributes
@@ -57,5 +57,3 @@ eng/common merge=ours
 eng/Common.props merge=ours
 eng/Versions.props merge=ours
 eng/Version.Details.xml merge=ours
-
-
