+++ src/Microsoft.Diagnostics.Monitoring.WebApi/Controllers/DiagController.cs
@@ -564,8 +564,9 @@ namespace Microsoft.Diagnostics.Monitoring.WebApi.Controllers
         }
 
         [HttpPost("parameters", Name = nameof(CaptureParameters))]
-        [ProducesWithProblemDetails(ContentTypes.ApplicationJson)]
+        [ProducesWithProblemDetails(ContentTypes.ApplicationJson, ContentTypes.TextPlain)]
         [ProducesResponseType(typeof(ProblemDetails), StatusCodes.Status429TooManyRequests)]
+        [ProducesResponseType(typeof(string), StatusCodes.Status200OK)]
         [ProducesResponseType(typeof(void), StatusCodes.Status202Accepted)]
         [EgressValidation]
         public async Task<ActionResult> CaptureParameters(
@@ -580,6 +581,8 @@ namespace Microsoft.Diagnostics.Monitoring.WebApi.Controllers
             [FromQuery]
             string name = null,
             [FromQuery]
+            string egressProvider = null,
+            [FromQuery]
             string tags = null)
         {
             if (!_parameterCapturingOptions.Value.GetEnabled())
@@ -590,59 +593,21 @@ namespace Microsoft.Diagnostics.Monitoring.WebApi.Controllers
             ProcessKey? processKey = Utilities.GetProcessKey(pid, uid, name);
             TimeSpan duration = Utilities.ConvertSecondsToTimeSpan(durationSeconds);
 
-            return await InvokeForProcess(async processInfo =>
-            {
-                return await InProcessResult(
-                    Utilities.ArtifactType_RequestParameters,
-                    processInfo,
-                    requestId => _captureParametersFactory.Create(requestId, processInfo.EndpointInfo, configuration, duration),
-                    tags);
-            }, processKey, Utilities.ArtifactType_RequestParameters);
-        }
-
-        [HttpGet("parameters", Name = nameof(GetCapturedParameters))]
-        [ProducesWithProblemDetails(ContentTypes.ApplicationJson, ContentTypes.TextPlain)]
-        [ProducesResponseType(typeof(ProblemDetails), StatusCodes.Status429TooManyRequests)]
-        [ProducesResponseType(typeof(string), StatusCodes.Status200OK)]
-        [ProducesResponseType(typeof(void), StatusCodes.Status202Accepted)]
-        [EgressValidation]
-        public Task<ActionResult> GetCapturedParameters(
-            [FromQuery]
-            int? pid = null,
-            [FromQuery]
-            Guid? uid = null,
-            [FromQuery]
-            string name = null,
-            [FromQuery]
-            Guid? requestId = null,
-            [FromQuery]
-            string egressProvider = null,
-            [FromQuery]
-            string tags = null)
-        {
-            if (!_parameterCapturingOptions.Value.GetEnabled())
-            {
-                return Task.FromResult<ActionResult>(this.FeatureNotEnabled(Strings.FeatureName_ParameterCapturing));
-            }
-
-            ProcessKey? processKey = Utilities.GetProcessKey(pid, uid, name);
-
-            return InvokeForProcess(processInfo =>
+            return await InvokeForProcess(processInfo =>
             {
                 CapturedParameterFormat format = ContentTypeUtilities.ComputeCapturedParameterFormat(Request.GetTypedHeaders().Accept) ?? CapturedParameterFormat.Json;
 
-                IArtifactOperation operation = _captureParametersFactory.CreateCapturedParameterFetcher(processInfo.EndpointInfo, requestId, format);
+                IArtifactOperation operation = _captureParametersFactory.Create(processInfo.EndpointInfo, configuration, duration, format);
 
                 return Result(
                     Utilities.ArtifactType_Parameters,
                     egressProvider,
                     operation,
                     processInfo,
-                    tags);
-
+                    tags,
+                    format != CapturedParameterFormat.PlainText);
             }, processKey, Utilities.ArtifactType_Parameters);
         }
-
 
         [HttpGet("stacks", Name = nameof(CaptureStacks))]
         [ProducesWithProblemDetails(ContentTypes.ApplicationJson, ContentTypes.TextPlain, ContentTypes.ApplicationSpeedscopeJson)]
@@ -651,15 +616,15 @@ namespace Microsoft.Diagnostics.Monitoring.WebApi.Controllers
         [ProducesResponseType(typeof(void), StatusCodes.Status202Accepted)]
         [EgressValidation]
         public Task<ActionResult> CaptureStacks(
-        [FromQuery]
+            [FromQuery]
             int? pid = null,
-        [FromQuery]
+            [FromQuery]
             Guid? uid = null,
-        [FromQuery]
+            [FromQuery]
             string name = null,
-        [FromQuery]
+            [FromQuery]
             string egressProvider = null,
-        [FromQuery]
+            [FromQuery]
             string tags = null)
         {
             if (!_callStacksOptions.Value.GetEnabled())
