+++ src/Microsoft.Diagnostics.Monitoring.WebApi/Utilities/Utilities.cs
@@ -17,7 +17,6 @@ namespace Microsoft.Diagnostics.Monitoring.WebApi
         public const string ArtifactType_Metrics = "livemetrics";
         public const string ArtifactType_Stacks = "stacks";
         public const string ArtifactType_Exceptions = "exceptions";
-        public const string ArtifactType_RequestParameters = "request_parameters";
         public const string ArtifactType_Parameters = "parameters";
 
         public static TimeSpan ConvertSecondsToTimeSpan(int durationSeconds)
